3 - LangChain/requirements.txt
openai==0.86.0
groq>=0.9.0
python-dotenv==1.2.2
langchain==0.3.25
langchain-openai==0.0.22
langchain-groq>=0.2.0
langchain-core==0.1.1
faiss-cpu==1.13.0
langchain-community==0.3.25
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
pypdf==5.6.0
sentence-transformers>=2.2.0